+++ CHANGES.md
@@ -54,6 +54,7 @@ Changes in 10.0
 - SQL remove use of nextval in rosariosis_fr.sql
 - Rename $pg_dumpPath configuration variable to $DatabaseDumpPath in config.inc.sample.php & DatabaseBackup.php
 - Build command for executing mysqldump in DatabaseBackup.php
+- SQL to extract Unix timestamp or epoch from date in Eligibility/Student.php, StudentList.php & TeacherCompletion.php
 
 Changes in 9.3
 --------------